[info] Simplification of language Restaurant classes (and their test). Extend from parent AbstractRestaurant.
