+++ src/AbstractRestaurant.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace FakerRestaurant;
+
+abstract class AbstractRestaurant extends \Faker\Provider\Base
+{
+    /**
+     * A random Food Name.
+     * @return string
+     */
+    public function foodName()
+    {
+        return static::randomElement(static::$foodNames);
+    }
+
+    /**
+     * A random Beverage Name.
+     * @return string
+     */
+    public function beverageName()
+    {
+        return static::randomElement(static::$beverageNames);
+    }
+
+    /**
+     * A random Dairy Name.
+     * @return string
+     */
+    public function dairyName()
+    {
+        return static::randomElement(static::$dairyNames);
+    }
+
+    /**
+     * A random Vegetable Name.
+     * @return string
+     */
+    public function vegetableName()
+    {
+        return static::randomElement(static::$vegetableNames);
+    }
+
+    /**
+     * A random Fruit Name.
+     * @return string
+     */
+    public function fruitName()
+    {
+        return static::randomElement(static::$fruitNames);
+    }
+
+    /**
+     * A random Meat Name.
+     * @return string
+     */
+    public function meatName()
+    {
+        return static::randomElement(static::$meatNames);
+    }
+
+    /**
+     * A random Sauce Name.
+     * @return string
+     */
+    public function sauceName()
+    {
+        return static::randomElement(static::$sauceNames);
+    }
+}
