+++ src/Provider/ar_SA/Restaurant.php
@@ -2,7 +2,9 @@
 
 namespace FakerRestaurant\Provider\ar_SA;
 
-class Restaurant extends \Faker\Provider\Base
+use FakerRestaurant\AbstractRestaurant;
+
+class Restaurant extends AbstractRestaurant
 {
     protected static $foodNames = [
         'بيتزا بالجبنة', 'همبرغر', 'برجر بالجبنة', 'برجر لحمبقري', 'برغر لحم بالجبنة',
@@ -102,67 +104,4 @@ class Restaurant extends \Faker\Provider\Base
         'صلصة الفلفل الحار',
         'صلصة الثوم',
     ];
-
-    /**
-     * A random Food Name.
-     * @return string
-     */
-    public function foodName()
-    {
-        return static::randomElement(static::$foodNames);
-    }
-
-    /**
-     * A random Beverage Name.
-     * @return string
-     */
-    public function beverageName()
-    {
-        return static::randomElement(static::$beverageNames);
-    }
-
-    /**
-     * A random Dairy Name.
-     * @return string
-     */
-    public function dairyName()
-    {
-        return static::randomElement(static::$dairyNames);
-    }
-
-    /**
-     * A random Vegetable Name.
-     * @return string
-     */
-    public function vegetableName()
-    {
-        return static::randomElement(static::$vegetableNames);
-    }
-
-    /**
-     * A random Fruit Name.
-     * @return string
-     */
-    public function fruitName()
-    {
-        return static::randomElement(static::$fruitNames);
-    }
-
-    /**
-     * A random Meat Name.
-     * @return string
-     */
-    public function meatName()
-    {
-        return static::randomElement(static::$meatNames);
-    }
-
-    /**
-     * A random Sauce Name.
-     * @return string
-     */
-    public function sauceName()
-    {
-        return static::randomElement(static::$sauceNames);
-    }
 }
