+++ src/Provider/fa_IR/Restaurant.php
@@ -2,9 +2,9 @@
 
 namespace FakerRestaurant\Provider\fa_IR;
 
-use Faker\Provider\Base;
+use FakerRestaurant\AbstractRestaurant;
 
-class Restaurant extends Base
+class Restaurant extends AbstractRestaurant
 {
     protected static $foodNames = [
         'پیتزا', 'همبرگر', 'چیز برگر', 'بیکن برگر', 'بیکن برگر با پنیر',
@@ -105,67 +105,4 @@ class Restaurant extends Base
         'سس تند چیلی',
         'سس سیر',
     ];
-
-    /**
-     * A random Food Name.
-     * @return string
-     */
-    public function foodName()
-    {
-        return static::randomElement(static::$foodNames);
-    }
-
-    /**
-     * A random Beverage Name.
-     * @return string
-     */
-    public function beverageName()
-    {
-        return static::randomElement(static::$beverageNames);
-    }
-
-    /**
-     * A random Dairy Name.
-     * @return string
-     */
-    public function dairyName()
-    {
-        return static::randomElement(static::$dairyNames);
-    }
-
-    /**
-     * A random Vegetable Name.
-     * @return string
-     */
-    public function vegetableName()
-    {
-        return static::randomElement(static::$vegetableNames);
-    }
-
-    /**
-     * A random Fruit Name.
-     * @return string
-     */
-    public function fruitName()
-    {
-        return static::randomElement(static::$fruitNames);
-    }
-
-    /**
-     * A random Meat Name.
-     * @return string
-     */
-    public function meatName()
-    {
-        return static::randomElement(static::$meatNames);
-    }
-
-    /**
-     * A random Sauce Name.
-     * @return string
-     */
-    public function sauceName()
-    {
-        return static::randomElement(static::$sauceNames);
-    }
 }
